# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Marketplace API",
    "description": "Marketplace API",
    "version": "1.0",
    "contact": {
      "name": "Marketplace",
      "url": "https://ordercloud.io",
      "email": "oheywood@four51.com"
    }
  },
  "servers": [
    {
      "url": "https://marketplace-api-qa.azurewebsites.net",
      "description": "Marketplace API"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "password": {
            "tokenUrl": "/oauth/token",
            "scopes": {
              "AddressAdmin": "",
              "AddressReader": "",
              "AdminAddressAdmin": "",
              "AdminAddressReader": "",
              "AdminUserAdmin": "",
              "AdminUserGroupAdmin": "",
              "AdminUserGroupReader": "",
              "AdminUserReader": "",
              "ApiClientAdmin": "",
              "ApiClientReader": "",
              "ApprovalRuleAdmin": "",
              "ApprovalRuleReader": "",
              "BuyerAdmin": "",
              "BuyerImpersonation": "",
              "BuyerReader": "",
              "BuyerUserAdmin": "",
              "BuyerUserReader": "",
              "CatalogAdmin": "",
              "CatalogReader": "",
              "CategoryAdmin": "",
              "CategoryReader": "",
              "CostCenterAdmin": "",
              "CostCenterReader": "",
              "CreditCardAdmin": "",
              "CreditCardReader": "",
              "FullAccess": "",
              "IncrementorAdmin": "",
              "IncrementorReader": "",
              "InventoryAdmin": "",
              "MeAddressAdmin": "",
              "MeAdmin": "",
              "MeCreditCardAdmin": "",
              "MessageConfigAssignmentAdmin": "",
              "MessageSenderAdmin": "",
...
    "schemas": {
      "BuyerAddressXP": {
        "type": "object",
        "example": {"Accessorials":["DestinationInsideDelivery"],"Email":""},
        "properties": {
          "Accessorials": {
            "type": "string",
            "enum": [
              "DestinationInsideDelivery",
              "DestinationLiftGate",
              "LimitedAccessDelivery",
              "ResidentialDelivery"
            ],
            "items": {
              "type": "string"
            }
          },
          "Email": {
            "type": "string"
          }
        }
      },
...
      "BuyerXp": {
        "type": "object",
        "example": {},
        "properties": {}
      },
...
      "CostCenterXp": {
        "type": "object",
        "example": {},
        "properties": {}
      },
...
      "MarketplaceAddressAssignment": {
        "type": "object",
        "example": {"ID":"","AddressID":"","UserID":"","UserGroupID":"","IsShipping":false,"IsBilling":false},
        "properties": {
          "ID": {
            "type": "string"
          },
          "AddressID": {
            "type": "string"
          },
          "UserID": {
            "type": "string"
          },
          "UserGroupID": {
            "type": "string"
          },
          "IsShipping": {
            "type": "boolean"
          },
          "IsBilling": {
            "type": "boolean"
          }
        }
      },
      "MarketplaceAddressBuyer": {
        "type": "object",
        "example": {"xp":{"Accessorials":["DestinationInsideDelivery"],"Email":""},"ID":"","DateCreated":"2018-01-01T00:00:00-06:00","CompanyName":"","FirstName":"","LastName":"","Street1":"","Street2":"","City":"","State":"","Zip":"","Country":"","Phone":"","AddressName":""},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BuyerAddressXP"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "DateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "CompanyName": {
            "type": "string"
          },
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "Street1": {
            "type": "string"
          },
          "Street2": {
            "type": "string"
          },
          "City": {
            "type": "string"
          },
          "State": {
            "type": "string"
          },
          "Zip": {
            "type": "string"
          },
          "Country": {
            "type": "string"
          },
          "Phone": {
            "type": "string"
          },
          "AddressName": {
            "type": "string"
          }
        }
      },
      "MarketplaceBuyer": {
        "type": "object",
        "example": {"xp":{},"ID":"","Name":"","DefaultCatalogID":"","Active":false},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BuyerXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "DefaultCatalogID": {
            "type": "string"
          },
          "Active": {
            "type": "boolean"
          }
        }
      },
...
      "MarketplaceCostCenter": {
        "type": "object",
        "example": {"xp":{},"ID":"","Name":"","Description":""},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CostCenterXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          }
        }
      },
...
      "MarketplaceUser": {
        "type": "object",
        "example": {"xp":{},"ID":"","Username":"","Password":"","FirstName":"","LastName":"","Email":"","Phone":"","TermsAccepted":"2018-01-01T00:00:00-06:00","Active":false,"AvailableRoles":[""],"DateCreated":"2018-01-01T00:00:00-06:00","PasswordLastSetDate":"2018-01-01T00:00:00-06:00"},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Username": {
            "type": "string"
          },
          "Password": {
            "type": "string",
            "format": "password"
          },
          "FirstName": {
            "type": "string"
          },
          "LastName": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          },
          "Phone": {
            "type": "string"
          },
          "TermsAccepted": {
            "type": "string",
            "format": "date-time"
          },
          "Active": {
            "type": "boolean"
          },
          "AvailableRoles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "DateCreated": {
            "type": "string",
            "format": "date-time"
          },
          "PasswordLastSetDate": {
            "type": "string",
            "format": "password"
          }
        }
      },
      "MarketplaceUserGroup": {
        "type": "object",
        "example": {"xp":{"Type":""},"ID":"","Name":"","Description":""},
        "properties": {
          "xp": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserGroupXp"
              }
            ]
          },
          "ID": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          },
          "Description": {
            "type": "string"
          }
        }
      },
      "MarketplaceUserGroupAssignment": {
        "type": "object",
        "example": {"ID":"","UserGroupID":"","UserID":""},
        "properties": {
          "ID": {
            "type": "string"
          },
          "UserGroupID": {
            "type": "string"
          },
          "UserID": {
            "type": "string"
          }
        }
      },
...
      "UserGroupXp": {
        "type": "object",
        "example": {"Type":""},
        "properties": {
          "Type": {
            "type": "string"
          }
        }
      },
      "UserXp": {
        "type": "object",
        "example": {},
        "properties": {}
      },
...
    }
  },
  "paths": {
...
    "/buyer": {
      "post": {
        "responses": {
          "201": {
            "description": "If ID is provided and an object with that ID already exists, a 409 (conflict) error is returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceBuyer"
                }
              }
            }
          }
        },
        "operationId": "Buyers.Create",
        "tags": [
          "Buyers"
        ],
        "requestBody": {"required":true,"description":"","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/MarketplaceBuyer"}]}}}},
        "summary": "POST Marketplace Buyer",
        "description": "If ID is provided and an object with that ID already exists, a 409 (conflict) error is returned.",
        "security": [
          {
            "OAuth2": [
              "BuyerAdmin",
              "FullAccess"
            ]
          }
        ]
      }
    },
...
    "/orchestration/{clientId}/{buyerId}/address": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceAddressBuyer"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostAddress",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace address buyer.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace address buyer.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Address",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/{buyerId}/addressassignment": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceAddressAssignment"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostAddressAssignment",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace address assignment.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace address assignment.","required":true,"schema":{"type":"string"}}],
        "summary": "POST AddressAssignment",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/{buyerId}/costcenter": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceCostCenter"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostCostCenter",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace cost center.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace cost center.","required":true,"schema":{"type":"string"}}],
        "summary": "POST CostCenter",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/{buyerId}/user": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceUser"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostUser",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace user.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace user.","required":true,"schema":{"type":"string"}}],
        "summary": "POST User",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/{buyerId}/usergroup": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceUserGroup"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostUserGroup",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace user group.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace user group.","required":true,"schema":{"type":"string"}}],
        "summary": "POST UserGroup",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/{buyerId}/usergroupassignment": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceUserGroupAssignment"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostUserGroupAssignment",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"buyerId","in":"path","description":"Buyer id of the marketplace user group assignment.","required":true,"schema":{"type":"string"}},{"name":"clientId","in":"path","description":"Client id of the marketplace user group assignment.","required":true,"schema":{"type":"string"}}],
        "summary": "POST UserGroupAssignment",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
    "/orchestration/{clientId}/buyer": {
      "post": {
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceBuyer"
                }
              }
            }
          }
        },
        "operationId": "OrchestrationUsers.PostBuyer",
        "tags": [
          "Orchestration Users"
        ],
        "parameters": [{"name":"clientId","in":"path","description":"Client id of the marketplace buyer.","required":true,"schema":{"type":"string"}}],
        "summary": "POST Buyer",
        "description": "",
        "security": [
          {
            "OAuth2": [
              "FullAccess"
            ]
          }
        ]
      }
    },
...
  }
}